chore: caches node modules in GitHub Actions workflow
Adds logic for caching the contents of the node_modules directory which speeds up workflow runs as long as the package.json file doesn't change (if the lock file is committed, it should be part of the caching key instead of the package.json file).

Adds pull requests as workflow triggers so people's contributions have tests running for them (for first-time contributors, a project maintainer must approve workflows to run, see https://github.blog/changelog/2021-04-22-github-actions-maintainers-must-approve-first-time-contributor-workflow-runs/).